Second commit
Changes in regustering a project: Now the managers are registered in arrays depending of the cuantity of managers from each side that the user inputs. Also changed the for loop to a while, so that the user is not asked to input another project after he has allready said no. Instead the while cile now closes also if the proyect counter reaches 10.

Changes in culminating a stage: Now the date is not inputed by the user, but taken from java Calendar.

Added the first version for publishing a capsule