Finished first submission
Added the utils package, containing the Utils class, which has commonly used methods, like validating inputs and printing method.

Capsule aproving: Changed the html generator to this method, as it was missplaced before. Also added the aprovation date of the capsule.

Added the deffault case, and deleated the part of the menu that is not needed for the first submission.